Remove misc unneeded modifications in libchrome.
Bug: 73270448
Test: Built locally. Treehugger.

(Reconstructed from 31babce11bb4f466b4754956e05cf62f48060fa4)

+++ base/sys_info_posix.cc
@@ -90,7 +90,10 @@ bool IsStatsZeroIfUnlimited(const base::FilePath& path) {
   if (HANDLE_EINTR(statfs(path.value().c_str(), &stats)) != 0)
     return false;
 
-  switch (stats.f_type) {
+  // In some platforms, |statfs_buf.f_type| is declared as signed, but some of
+  // the values will overflow it, causing narrowing warnings. Cast to the
+  // largest possible unsigned integer type to avoid it.
+  switch (static_cast<uintmax_t>(stats.f_type)) {
     case TMPFS_MAGIC:
     case HUGETLBFS_MAGIC:
     case RAMFS_MAGIC:
